+++ src/Fields/Button.php
@@ -108,7 +108,7 @@ class Button extends Field_Base {
 	 *
 	 * @return int
 	 */
-	public function sanitize_callback( mixed $value ): int {
+	public function default_sanitize_callback( mixed $value ): int {
 		// bail if value is null.
 		if ( is_null( $value ) ) {
 			return 0;