+++ src/Fields/PermalinkSlug.php
@@ -127,7 +127,7 @@ class PermalinkSlug extends Field_Base {
 	 *
 	 * @return string
 	 */
-	public function sanitize_callback( mixed $value ): string {
+	public function default_sanitize_callback( mixed $value ): string {
 		// bail if value is null.
 		if ( is_null( $value ) ) {
 			return '';