+++ src/Fields/Radio.php
@@ -112,7 +112,7 @@ class Radio extends Field_Base {
 	 *
 	 * @return int
 	 */
-	public function sanitize_callback( mixed $value ): int {
+	public function default_sanitize_callback( mixed $value ): int {
 		// bail if value is null.
 		if ( is_null( $value ) ) {
 			return 0;