+++ src/Fields/Select.php
@@ -103,4 +103,15 @@ class Select extends Field_Base {
 	public function set_options( array $options ): void {
 		$this->options = $options;
 	}
+
+	/**
+	 * The sanitize callback for this field.
+	 *
+	 * @param mixed $value The value to save.
+	 *
+	 * @return mixed
+	 */
+	public function default_sanitize_callback( mixed $value ): mixed {
+		return $value;
+	}
 }