+++ src/Import.php
@@ -83,6 +83,11 @@ class Import extends Base_Object {
 		// check nonce.
 		check_ajax_referer( 'settings-import', 'nonce' );
 
+		// bail if user has not the capability to change settings.
+		if( ! current_user_can( $this->get_settings_obj()->get_capability() ) ) {
+			wp_send_json_error();
+		}
+
 		// get the translations.
 		$translations = $this->get_settings_obj()->get_translations();